Mail Fix (#328)
* Исправления писем.

Исправлен DoAfter вскрытия писем (теперь он происходит). Проблему вызвало возможность молей есть письма (DoAfter не успевал вызываться до попытки съесть).

Исправлен баг не позволяющий открывать письма без получателей. Теперь они спавнятся без получателя и исправно открываются.

+++ Content.Server/_NF/Mail/DelayedItemSystem.cs
@@ -25,13 +25,18 @@ namespace Content.Server.Mail
             SubscribeLocalEvent<DelayedItemComponent, EntGotRemovedFromContainerMessage>(OnRemovedFromContainer);
         }
 
+        // Reserve edit start: mail-fix #328
         /// <summary>
         /// EntGotRemovedFromContainerMessage handler - spawn the intended entity after removed from a container.
         /// </summary>
-        private void OnRemovedFromContainer(EntityUid uid, DelayedItemComponent component, ContainerModifiedMessage args)
+        private void OnRemovedFromContainer(EntityUid uid, DelayedItemComponent component, EntGotRemovedFromContainerMessage args)
         {
+            if (TerminatingOrDeleted(uid))
+                return;
+
             Spawn(component.Item, Transform(uid).Coordinates);
         }
+        // Reserve edit end: mail-fix #328
 
         /// <summary>
         /// GotEquippedHandEvent handler - destroy the placeholder.
@@ -49,13 +54,18 @@ namespace Content.Server.Mail
             EntityManager.DeleteEntity(uid);
         }
 
+        // Reserve edit start: mail-fix #328
         /// <summary>
         /// OnDamageChanged handler - item has taken damage (e.g. inside the envelope), spawn the intended entity outside of any container and delete the placeholder.
         /// </summary>
         private void OnDamageChanged(EntityUid uid, DelayedItemComponent component, DamageChangedEvent args)
         {
+            if (TerminatingOrDeleted(uid))
+                return;
+
             Spawn(component.Item, Transform(uid).Coordinates);
             EntityManager.DeleteEntity(uid);
         }
+        // Reserve edit end: mail-fix #328
     }
 }
